requirements.txt
pandas
openpyxl
psycopg2-binary
python-dotenv
requests